+++ AllBirds.Application/Services/ProductColorServices/ProductColorService.cs
@@ -1,5 +1,6 @@
 ﻿using AllBirds.Application.Contracts;
 using AllBirds.Application.Services.ProductColorImageServices;
+using AllBirds.DTOs.CategoryDTOs;
 using AllBirds.DTOs.OrderStateDTOs;
 using AllBirds.DTOs.ProductColorDTOs;
 using AllBirds.DTOs.ProductColorImageDTOs;
@@ -29,7 +30,7 @@ namespace AllBirds.Application.Services.ProductColorServices
             productColotImageService = _productColotImageService;
 
         }
-        public async Task<ResultView<CreateProductColorDTO>> CreateAsync(CreateProductColorDTO productColorDTO,string ImagePath)
+        public async Task<ResultView<CreateProductColorDTO>> CreateAsync(CreateProductColorDTO productColorDTO, string ImagePath)
         {
             ResultView<CreateProductColorDTO> result = new();
             try
@@ -39,7 +40,7 @@ namespace AllBirds.Application.Services.ProductColorServices
                 {
                     result.IsSuccess = false;
                     result.Data = null;
-                    result.Msg = "Product Color  Is ALready Exist";
+                    result.Msg = "Product Color Is ALready Exist";
                     return result;
                 }
 
@@ -50,29 +51,29 @@ namespace AllBirds.Application.Services.ProductColorServices
                     prCL.Images = new List<ProductColorImage>();
                 }
 
+                if (!Directory.Exists(ImagePath))
+                {
+                    Directory.CreateDirectory(ImagePath);
+                }
+                string? mainImageUniqueFileName = null;
                 foreach (IFormFile formFile in productColorDTO.Images)
                 {
-
-                    if (!Directory.Exists(ImagePath))
-                    {
-                        Directory.CreateDirectory(ImagePath);
-                    }
-
                     string uniqueFileName = Guid.NewGuid().ToString() + "_" + formFile.FileName;
                     string filePath = Path.Combine(ImagePath, uniqueFileName);
-
                     using (var fileStream = new FileStream(filePath, FileMode.Create))
                     {
                         await formFile.CopyToAsync(fileStream);
                     }
-
-                    ProductColorImage productColorImage = new() { ImagePath = "/Images/" + "/ProductColorImages/" + uniqueFileName   };
-
+                    ProductColorImage productColorImage = new() { ImagePath = "/images/" + "/product-color-images/" + uniqueFileName };
+                    if (mainImageUniqueFileName is null)
+                        mainImageUniqueFileName = productColorImage.ImagePath;
                     prCL.Images.Add(productColorImage);
-
                 }
                 ProductColor created = await productColorRepository.CreateAsync(prCL);
                 await productColorRepository.SaveChangesAsync();
+                created.MainImageId = created.Images.FirstOrDefault(i => i.ImagePath == mainImageUniqueFileName).Id;
+                await productColorRepository.SaveChangesAsync();
+
 
                 result.IsSuccess = true;
                 result.Data = mapper.Map<CreateProductColorDTO>(created);
@@ -125,27 +126,32 @@ namespace AllBirds.Application.Services.ProductColorServices
                 result.IsSuccess = false;
                 result.Data = null;
                 result.Msg = "Error Happened While Creating With Ex :" + ex.Message;
-
             }
-
             return result;
-
         }
+
         public async Task<ResultView<List<GetALlProductColorDTO>>> GetAllAsync(int Id)
         {
             ResultView<List<GetALlProductColorDTO>> result = new();
             try
             {
-                var item = (await productColorRepository.GetAllAsync()).Where(b => !b.IsDeleted && b.ProductId == Id)
-                            .Include(b => b.Product)
-                            .Include(s => s.Color)
-                            .Include(r => r.Images);
+                var item = (await productColorRepository.GetAllAsync()).Where(b => b.ProductId == Id).Select(P => new GetALlProductColorDTO
+                {
+                    PNameEn = P.Product.NameEn,
+                    PNameAr = P.Product.NameAr,
+                    ProductNo = P.Product.ProductNo,
+                    Id = P.Id,
+                    ColorNameAr = P.Color.NameAr,
+                    ColorNameEn = P.Color.NameEn,
+                    ColorCode = P.Color.Code,
+                    MainImageId = P.MainImageId,
+                    IsDeleted = P.IsDeleted,
+                    MainImagePath = P.Images.FirstOrDefault(I => I.Id == P.MainImageId).ImagePath
+                }).ToList();
                 if (item.Count() != 0)
                 {
-
                     result.IsSuccess = true;
                     result.Data = mapper.Map<List<GetALlProductColorDTO>>(item);
-
                     result.Msg = "Get All Product's Colors successfully";
                 }
                 else
@@ -154,21 +160,59 @@ namespace AllBirds.Application.Services.ProductColorServices
                     result.Data = null;
                     result.Msg = "Product's Color's list Is Empty";
                 }
-
             }
             catch (Exception ex)
             {
-
                 result.IsSuccess = false;
                 result.Data = null;
                 result.Msg = "Error Happened While Getting All Product Colors With Ex :" + ex.Message;
-
-
             }
-
             return result;
-
-
+        }
+        
+        public async Task<ResultView<EntityPaginated<GetALlProductColorDTO>>> GetAllPaginatedAsync(int Id, int pageNumber, int pageSize)
+        {
+            ResultView<EntityPaginated<GetALlProductColorDTO>> result = new();
+            try
+            {
+                var item = (await productColorRepository.GetAllAsync()).Where(b => b.ProductId == Id).Select(P => new GetALlProductColorDTO
+                {
+                    PNameEn = P.Product.NameEn,
+                    PNameAr = P.Product.NameAr,
+                    ProductNo = P.Product.ProductNo,
+                    Id = P.Id,
+                    ColorNameAr = P.Color.NameAr,
+                    ColorNameEn = P.Color.NameEn,
+                    ColorCode = P.Color.Code,
+                    MainImageId = P.MainImageId,
+                    IsDeleted = P.IsDeleted,
+                    MainImagePath = P.Images.FirstOrDefault(I => I.Id == P.MainImageId).ImagePath
+                }).Skip((pageNumber - 1) * pageSize).Take(pageSize).ToList();
+                int totalPrdColors = (await productColorRepository.GetAllAsync()).Count(pc => pc.ProductId == Id);
+                if (item.Count > 0)
+                {
+                    result.IsSuccess = true;
+                    result.Data = new EntityPaginated<GetALlProductColorDTO>
+                    {
+                        Data = mapper.Map<List<GetALlProductColorDTO>>(item),
+                        Count = totalPrdColors
+                    };
+                    result.Msg = "Get All Product's Colors successfully";
+                }
+                else
+                {
+                    result.IsSuccess = false;
+                    result.Data = null;
+                    result.Msg = "Product's Color's list Is Empty";
+                }
+            }
+            catch (Exception ex)
+            {
+                result.IsSuccess = false;
+                result.Data = null;
+                result.Msg = "Error Happened While Getting All Product Colors With Ex :" + ex.Message;
+            }
+            return result;
         }
 
         public async Task<ResultView<List<GetOneProductColorDTO>>> GetAllWithDeletedAsync()
@@ -211,45 +255,34 @@ namespace AllBirds.Application.Services.ProductColorServices
 
         public async Task<ResultView<GetOneProductColorDTO>> GetByIdAsync(int id)
         {
-
-
             ResultView<GetOneProductColorDTO> result = new();
             try
             {
-                var item = (await productColorRepository.GetAllAsync())
-                    .Where(b => !b.IsDeleted && b.Id == id)
+                ProductColor? item = (await productColorRepository.GetAllAsync())
+                    .Where(b => b.Id == id)
                     .Include(b => b.Product)
                     .Include(s => s.Color)
                     .Include(r => r.Images)
-                    .Include(s => s.AvailableSizes).ThenInclude(a => a.Size);
-
-                if (item != null)
+                    .FirstOrDefault();
+                if (item is not null)
                 {
-
-
-
                     result.IsSuccess = true;
                     result.Data = mapper.Map<GetOneProductColorDTO>(item);
-                    result.Msg = "Get  Product's Color successfully";
+                    result.Msg = "Product Color Images Fetched successfully";
                 }
                 else
                 {
                     result.IsSuccess = false;
                     result.Data = null;
-                    result.Msg = "Product's Color Not Found";
+                    result.Msg = "Product Color Images Not Found";
                 }
-
             }
             catch (Exception ex)
             {
-
                 result.IsSuccess = false;
                 result.Data = null;
-                result.Msg = "Error Happened While Getting  Product Color By Id With Ex :" + ex.Message;
-
-
+                result.Msg = $"Error Happened While Getting Product Color Images, {ex.Message}";
             }
-
             return result;
         }
 
@@ -259,16 +292,53 @@ namespace AllBirds.Application.Services.ProductColorServices
             ResultView<GetOneProductColorDTO> result = new();
             try
             {
-                var item = (await productColorRepository.GetAllAsync()).FirstOrDefault(b => b.Id == id);
+                ProductColor item = (await productColorRepository.GetAllAsync()).FirstOrDefault(b => b.Id == id);
                 if (item != null)
                 {
+                    if (item.IsDeleted)
+                    {
 
-                    var deleted = await productColorRepository.DeleteAsync(item);
-                    await productColorRepository.SaveChangesAsync();
+                        bool CheckImage = (await productColorRepository.GetAllAsync()).AsNoTracking().Any(P => P.Images.Any(i => i.ProductColorId == id));
+                        bool CheckSize = (await productColorRepository.GetAllAsync()).AsNoTracking().Any(P => P.AvailableSizes.Any(pcs => pcs.ProductColorId == id));
+                        if (CheckImage && CheckSize)
+                        {
+                            result.Data = mapper.Map<GetOneProductColorDTO>(item);
+                            result.IsSuccess = false;
+                            result.Msg = $"Cannot Delete This Color As There Are Images & Sizes That Depend On It";
+                        }
+                        else if (CheckSize)
+                        {
+                            result.Data = mapper.Map<GetOneProductColorDTO>(item);
+                            result.IsSuccess = false;
+                            result.Msg = $"This Product Color Have Depend On It Like Size..! Sorry Cannot Delete Thais Color ";
+                        }
+                        else if (CheckImage)
+                        {
+                            result.Data = mapper.Map<GetOneProductColorDTO>(item);
+                            result.IsSuccess = false;
+                            result.Msg = $"This Product Color Have Depend On It Like Image..! Sorry Cannot Delete Thais Color ";
+                        }
+                        else
+                        {
+                            ProductColor deleted = await productColorRepository.DeleteAsync(item);
+                            await productColorRepository.SaveChangesAsync();
+                            
+                            result.IsSuccess = true;
+                            result.Data = mapper.Map<GetOneProductColorDTO>(deleted);
+                            result.Msg = "Deleted Product Color Successfully";
 
-                    result.IsSuccess = true;
-                    result.Data = mapper.Map<GetOneProductColorDTO>(deleted);
-                    result.Msg = "Deleting Product's Color successfully";
+                        }
+                    }
+                    else
+                    {
+                        item.IsDeleted = true;
+                        int saveStatus = await productColorRepository.SaveChangesAsync();
+                        //if (saveStatus > 0)
+                        //{
+                        result.IsSuccess = true;
+                        result.Data = mapper.Map<GetOneProductColorDTO>(item);
+                        result.Msg = $"Product Color Soft Deleted Successfully";
+                    }
                 }
                 else
                 {
@@ -291,42 +361,42 @@ namespace AllBirds.Application.Services.ProductColorServices
             return result;
         }
 
-        public async Task<ResultView<GetOneProductColorDTO>> SoftDeleteAsync(int id)
-        {
-            ResultView<GetOneProductColorDTO> result = new();
-            try
-            {
-                var item = (await productColorRepository.GetAllAsync()).FirstOrDefault(b => b.Id == id);
-                if (item != null)
-                {
+        //public async Task<ResultView<GetOneProductColorDTO>> SoftDeleteAsync(int id)
+        //{
+        //    ResultView<GetOneProductColorDTO> result = new();
+        //    try
+        //    {
+        //        var item = (await productColorRepository.GetAllAsync()).FirstOrDefault(b => b.Id == id);
+        //        if (item != null)
+        //        {
 
-                    item.IsDeleted = true;
-                    await productColorRepository.SaveChangesAsync();
+        //            item.IsDeleted = true;
+        //            await productColorRepository.SaveChangesAsync();
 
-                    result.IsSuccess = true;
-                    result.Data = mapper.Map<GetOneProductColorDTO>(item);
-                    result.Msg = "Soft Deleting Product's Color successfully";
-                }
-                else
-                {
-                    result.IsSuccess = false;
-                    result.Data = null;
-                    result.Msg = "Product's Color Not Found";
-                }
+        //            result.IsSuccess = true;
+        //            result.Data = mapper.Map<GetOneProductColorDTO>(item);
+        //            result.Msg = "Soft Deleting Product's Color successfully";
+        //        }
+        //        else
+        //        {
+        //            result.IsSuccess = false;
+        //            result.Data = null;
+        //            result.Msg = "Product's Color Not Found";
+        //        }
 
-            }
-            catch (Exception ex)
-            {
+        //    }
+        //    catch (Exception ex)
+        //    {
 
-                result.IsSuccess = false;
-                result.Data = null;
-                result.Msg = "Error Happened While Delete Product Color By Id With Ex :" + ex.Message;
+        //        result.IsSuccess = false;
+        //        result.Data = null;
+        //        result.Msg = "Error Happened While Delete Product Color By Id With Ex :" + ex.Message;
 
 
-            }
+        //    }
 
-            return result;
-        }
+        //    return result;
+        //}
 
 
     }
